Functional WebClient and Server Waiting Room
Created the ability to make and add new players to game lobbies and send new player info to host as well as display the lobby in the webclient.

For the purposes of testing, use "servertest" through the WebClient in order for the WebClient to fake being a host.

Other than that, check IP addresses and port numbers in the Server/james_server.js and Website/main.js files.

TODO:
- Remove client from host's client list when client connection gets closed
- Close client connections when host connection gets closed
- Add a check for duplicate usernames in the same host/lobby